Simplified the implementation. Removed the "utils" file.
